backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
chromadb==0.4.22
sentence-transformers==2.5.1
torch==2.2.0
transformers==4.37.2 